Add function to define parameters for single file
Modifies the code to set exposure time and airmass
values for each files from the config file if they are present.
This allows for more precise data reduction and flux calibration when specific values are needed for individual files, while falling back to the default values if not specified.

This is a example setting:
```json
"science" : {
        "skip_science": false,
        "science_dir": "data/2025-01-05/sci",
        "files":{
            "ALIa050131.fits": {
                "exptime": 750,
                "airmass": 1.19594924806758
            },
            "ALIa050132.fits": {
                "exptime": 750,
                "airmass": 1.16128879097752
            }
        },
        "exptime": 1000,
        "airmass": 1.17
    }
```